Moving feng functionality to own module
